requirements.txt
streamlit
ollama
langchain-community
langchain-google-vertexai
langchain
python-dotenv